# src/main/resources/static/openapi.yaml
openapi: 3.0.0
info:
  title: HARCREJ
  description: API for managing the Polish Scouts organization.
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
...
  /api/v1/szczep:
    post:
      tags:
        - szczep
      summary: Create a new Szczep
      operationId: postSzczep
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Szczep'
      responses:
        '201':
          description: Szczep created successfully
    get:
      tags:
        - szczep
      summary: Get a list of all Szczep
      operationId: getListOfSzczep
      responses:
        '200':
          description: A list of Szczep
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Szczep'
  /api/v1/szczep/{szczepId}:
    get:
      tags:
        - szczep
      summary: Get details of a specific Szczep
      operationId: getSzczepById
      parameters:
        - in: path
          name: szczepId
          schema:
            type: integer
          required: true
      responses:
        '200':
          description: Szczep details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Szczep'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - szczep
      summary: Delete a specific Szczep
      operationId: deleteSzczep
      parameters:
        - in: path
          name: szczepId
          schema:
            type: integer
          required: true
      responses:
        '204':
          description: Szczep deleted successfully
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      summary: Update a Szczep
      operationId: updateSzczep
      tags:
        - szczep
      parameters:
        - in: path
          name: szczepId
          schema:
            type: integer
          required: true
      requestBody:
        description: Updated Szczep
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Szczep"
      responses:
        "200":
          description: Szczep updated successfully
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/szczep/{szczepId}/druzyna/{druzynaId}:
    post:
      tags:
        - Szczep
      summary: Link a Drużyna to a Szczep
      description: Associates a specific Drużyna with a Szczep.
      operationId: linkDruzynaToSzczep
      parameters:
        - name: szczepId
          in: path
          required: true
          schema:
            type: integer
          description: The unique ID of the Szczep to which the Drużyna will be linked.
        - name: druzynaId
          in: path
          required: true
          schema:
            type: integer
          description: The unique ID of the Drużyna to be linked to the Szczep.
      responses:
        '200':
          description: Drużyna successfully linked to Szczep.
...
components:
  schemas:
...
    Address:
      type: object
      properties:
        addressLine1:
          type: string
          maxLength: 30
        addressLine2:
          type: string
          maxLength: 30
        city:
          type: string
          maxLength: 30
        postcode:
          type: string
          maxLength: 10
...
    Stats:
      type: object
      properties:
        totalParticipants:
          type: integer
        totalInstruktorzy:
          type: integer
        totalZbiorki:
          type: integer
        totalWycieczki:
          type: integer
...
    Szczep:
      type: object
      properties:
        name:
          type: string
        areaOfOperation:
          type: string
        szczepowy:
          type: object
          $ref: '#/components/schemas/User'
        druzyny:
          type: array
          items:
            $ref: '#/components/schemas/Druzyna'
    Druzyna:
      type: object
      properties:
        name:
          type: string
        addressOfOperation:
          type: object
          $ref: '#/components/schemas/Address'
        druzynowy:
          type: object
          $ref: '#/components/schemas/User'
        type:
          type: string
          enum: [ harcerze, wedrownicy, zuchy ]
        komenda:
          type: array
          items:
            $ref: '#/components/schemas/User'
        currentStats:
          type: object
          $ref: '#/components/schemas/Stats'
        yearEstablished:
          type: string
    User:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        mobile:
          type: string
        dateOfBirth:
          type: string
          format: date
        address:
          type: object
          $ref: '#/components/schemas/Address'
        stopien:
          type: string
...
    ErrorResponse:
      required:
        - message
        - errors
      type: object
      properties:
        message:
          type: string
          example: Bad request
        errors:
          type: array
          description: Validation/ error messages
          items:
            type: string
            example: Szczep id not found
  responses:
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...